chore(regen): regenerate spy-sourced stdlib and spy-test C# for this round's compiler changes
Required by the pre-push staleness gate, and the diff is a finding channel rather than a formality
(#1354's lesson). What moved, each traceable to a change in this round:

- `(int)(value * 1000.0)` -> `NumericCheckedCast.ToInt(...)` in Socket: the `as!` narrowing that used
  to wrap silently is now checked (#1306).
- `_Random.Next(1 << k)` -> `1 << (int)(k)`: the shift count converts to int (#1315).
- `switch (r)` -> `switch (((double)r!))`: a match subject now carries its narrowed read (#1299).
- `x.Rstrip()` -> `StringExtensions.Rstrip(x)` and `Split(...)` wrapped in `new Sharpy.List<string>`:
  static-extension dispatch and #1251 sequence materialization, both pre-existing and only visible
  now because these artifacts had gone stale.

+++ src/Sharpy.Stdlib/Random/Random.cs
@@ -167,7 +167,7 @@ namespace Sharpy
                 throw new global::Sharpy.ValueError("number of bits must be <= 30 for int return");
             }
 
-            return _Random.Next(1 << k);
+            return _Random.Next(1 << (int)(k));
         }
 
         /// <summary>
